HEEDLS-572 Move accessibility test fixtures into folder; cleanup test files; remove seleniumtestcollection
