--- DigitalLearningSolutions.Web.AutomatedUiTests/AccessibilityTests/SeleniumTestCollection.cs
@@ -1,9 +1,0 @@
-﻿namespace DigitalLearningSolutions.Web.AutomatedUiTests.AccessibilityTests
-{
-    using Xunit;
-
-    [CollectionDefinition("Selenium test collection")]
-    public class SeleniumTestCollection
-    {
-    }
-}
