--- DigitalLearningSolutions.Web.AutomatedUiTests/AccessibilityTestsFixture.cs
@@ -1,25 +1,0 @@
-namespace DigitalLearningSolutions.Web.AutomatedUiTests
-{
-    using System;
-    using DigitalLearningSolutions.Web.AutomatedUiTests.TestHelpers;
-    using OpenQA.Selenium;
-
-    public class AccessibilityTestsFixture<TStartup> : IDisposable where TStartup : class
-    {
-        internal readonly string BaseUrl;
-        internal readonly IWebDriver Driver;
-
-        public AccessibilityTestsFixture()
-        {
-            var factory = new SeleniumServerFactory<TStartup>();
-            BaseUrl = factory.RootUri;
-            Driver = DriverHelper.CreateHeadlessChromeDriver();
-        }
-
-        public void Dispose()
-        {
-            Driver.Quit();
-            Driver.Dispose();
-        }
-    }
-}
